@@ -1,181 +1,198 @@
 import 'dart:io';
 import 'dart:typed_data';
-import 'package:firebase_core/firebase_core.dart';
+import 'dart:convert';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:injectable/injectable.dart';
-import 'package:universal_html/html.dart' as html;
 import '../../domain/models/pdf_document.dart';
 import '../../domain/models/pdf_bookmark.dart';
-import 'package:pdf_learner_v2/domain/models/user_model.dart';
-import 'package:pdf_learner_v2/domain/models/pdf_document.dart';
-import 'package:pdf_learner_v2/domain/models/pdf_bookmark.dart';
 
-@Injectable()
 class FirebaseService {
   final FirebaseAuth _auth;
   final FirebaseFirestore _firestore;
   final FirebaseStorage _storage;
 
-  FirebaseService(this._auth, this._firestore, this._storage);
+  FirebaseService(
+    this._auth,
+    this._firestore,
+    this._storage,
+  );
 
-  // 초기화
-  Future<void> initialize() async {
-    await Firebase.initializeApp();
-  }
-
-  // 인증 관련 메서드
+  // 유저 상태 관련 getter
   Stream<User?> get authStateChanges => _auth.authStateChanges();
   User? get currentUser => _auth.currentUser;
+  bool get isLoggedIn => currentUser != null;
+  bool get isAnonymous => currentUser?.isAnonymous ?? true;
+  String? get userId => currentUser?.uid;
+
+  // Firestore 컬렉션 참조
+  CollectionReference<Map<String, dynamic>> get usersCollection => _firestore.collection('users');
   
-  Future<UserCredential?> signInWithEmailAndPassword(String email, String password) async {
+  CollectionReference<Map<String, dynamic>> getUserDocumentsCollection(String userId) {
+    return usersCollection.doc(userId).collection('documents');
+  }
+  
+  CollectionReference<Map<String, dynamic>> getUserBookmarksCollection(String userId) {
+    return usersCollection.doc(userId).collection('bookmarks');
+  }
+  
+  // 스토리지 참조
+  Reference getUserStorageRef(String userId) {
+    return _storage.ref().child('users/$userId');
+  }
+  
+  Reference getPDFStorageRef(String userId, String documentId) {
+    return getUserStorageRef(userId).child('pdfs/$documentId.pdf');
+  }
+  
+  // 파일 업로드
+  Future<String> uploadFile(String userId, String path, File file) async {
     try {
-      return await _auth.signInWithEmailAndPassword(
-        email: email,
-        password: password,
-      );
+      final ref = _storage.ref().child('users/$userId/$path');
+      final uploadTask = await ref.putFile(file);
+      return await uploadTask.ref.getDownloadURL();
     } catch (e) {
       rethrow;
     }
   }
-
-  Future<UserCredential?> createUserWithEmailAndPassword(String email, String password) async {
+  
+  // 웹용 파일 업로드
+  Future<String> uploadBytes(String userId, String path, Uint8List bytes, String contentType) async {
     try {
-      return await _auth.createUserWithEmailAndPassword(
-        email: email,
-        password: password,
-      );
+      final ref = _storage.ref().child('users/$userId/$path');
+      final metadata = SettableMetadata(contentType: contentType);
+      final uploadTask = await ref.putData(bytes, metadata);
+      return await uploadTask.ref.getDownloadURL();
     } catch (e) {
       rethrow;
     }
   }
-
-  Future<void> signOut() async {
-    await _auth.signOut();
-  }
-
-  Future<void> sendPasswordResetEmail(String email) async {
-    await _auth.sendPasswordResetEmail(email: email);
-  }
-
-  // Firestore 관련 메서드
-  Future<void> createUser(UserModel user) async {
-    await _firestore.collection('users').doc(user.id).set(user.toJson());
-  }
-
-  Future<void> updateUser(UserModel user) async {
-    await _firestore.collection('users').doc(user.id).update(user.toJson());
-  }
-
-  Future<void> deleteUser(String userId) async {
-    await _firestore.collection('users').doc(userId).delete();
-  }
-
-  Future<UserModel?> getUser(String userId) async {
-    final doc = await _firestore.collection('users').doc(userId).get();
-    if (doc.exists) {
-      return UserModel.fromJson(doc.data()!);
+  
+  // 파일 다운로드 URL 획득
+  Future<String> getDownloadURL(String path) async {
+    try {
+      return await _storage.ref().child(path).getDownloadURL();
+    } catch (e) {
+      rethrow;
     }
-    return null;
+  }
+  
+  // 파일 삭제
+  Future<void> deleteFile(String path) async {
+    try {
+      await _storage.ref().child(path).delete();
+    } catch (e) {
+      rethrow;
+    }
   }
 
   // PDF 문서 관련 메서드
-  Future<void> createDocument(PDFDocument document) async {
-    await _firestore.collection('documents').doc(document.id).set(document.toJson());
+  Future<List<PDFDocument>> getPDFDocuments() async {
+    final user = _auth.currentUser;
+    if (user == null) return [];
+    
+    final snapshot = await getUserDocumentsCollection(user.uid).get();
+    return snapshot.docs.map((doc) => PDFDocument.fromMap(doc.data())).toList();
   }
-
-  Future<void> updateDocument(PDFDocument document) async {
-    await _firestore.collection('documents').doc(document.id).update(document.toJson());
+  
+  Future<PDFDocument?> getPDFDocument(String id) async {
+    final user = _auth.currentUser;
+    if (user == null) return null;
+    
+    final doc = await getUserDocumentsCollection(user.uid).doc(id).get();
+    if (!doc.exists) return null;
+    
+    return PDFDocument.fromMap(doc.data()!);
   }
-
-  Future<void> deleteDocument(String documentId) async {
-    await _firestore.collection('documents').doc(documentId).delete();
+  
+  Future<void> addPDFDocument(PDFDocument document) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserDocumentsCollection(user.uid).doc(document.id).set(document.toMap());
   }
-
-  Future<PDFDocument?> getDocument(String documentId) async {
-    final doc = await _firestore.collection('documents').doc(documentId).get();
-    if (doc.exists) {
-      return PDFDocument.fromJson(doc.data()!);
-    }
-    return null;
+  
+  Future<void> updatePDFDocument(PDFDocument document) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserDocumentsCollection(user.uid).doc(document.id).update(document.toMap());
   }
-
-  Stream<List<PDFDocument>> getDocuments() {
-    return _firestore
-        .collection('documents')
-        .snapshots()
-        .map((snapshot) => snapshot.docs
-            .map((doc) => PDFDocument.fromJson(doc.data()))
-            .toList());
+  
+  Future<void> deletePDFDocument(String id) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserDocumentsCollection(user.uid).doc(id).delete();
   }
-
+  
   // 북마크 관련 메서드
-  Future<void> createBookmark(PDFBookmark bookmark) async {
-    await _firestore.collection('bookmarks').doc(bookmark.id).set(bookmark.toJson());
-  }
-
-  Future<void> updateBookmark(PDFBookmark bookmark) async {
-    await _firestore.collection('bookmarks').doc(bookmark.id).update(bookmark.toJson());
-  }
-
-  Future<void> deleteBookmark(String bookmarkId) async {
-    await _firestore.collection('bookmarks').doc(bookmarkId).delete();
-  }
-
-  Future<PDFBookmark?> getBookmark(String bookmarkId) async {
-    final doc = await _firestore.collection('bookmarks').doc(bookmarkId).get();
-    if (doc.exists) {
-      return PDFBookmark.fromJson(doc.data()!);
-    }
-    return null;
-  }
-
-  Stream<List<PDFBookmark>> getBookmarks(String documentId) {
-    return _firestore
-        .collection('bookmarks')
+  Future<List<PDFBookmark>> getBookmarks(String documentId) async {
+    final user = _auth.currentUser;
+    if (user == null) return [];
+    
+    final snapshot = await getUserBookmarksCollection(user.uid)
         .where('documentId', isEqualTo: documentId)
-        .snapshots()
-        .map((snapshot) => snapshot.docs
-            .map((doc) => PDFBookmark.fromJson(doc.data()))
-            .toList());
+        .get();
+    return snapshot.docs.map((doc) => PDFBookmark.fromMap(doc.data())).toList();
   }
-
-  // Storage 관련 메서드
-  Future<String> uploadPDFFile(String filePath, String fileName, {Uint8List? bytes}) async {
-    final ref = _storage.ref().child('pdfs/$fileName');
+  
+  Future<PDFBookmark?> getBookmark(String documentId, String bookmarkId) async {
+    final user = _auth.currentUser;
+    if (user == null) return null;
     
-    if (bytes != null) {
-      // 웹 환경인 경우 바이트 데이터 업로드
-      await ref.putData(bytes);
-    } else {
-      // 네이티브 환경인 경우 파일 업로드
-      final file = File(filePath);
-      await ref.putFile(file);
-    }
+    final doc = await getUserBookmarksCollection(user.uid).doc(bookmarkId).get();
+    if (!doc.exists) return null;
     
+    return PDFBookmark.fromMap(doc.data()!);
+  }
+  
+  Future<void> addBookmark(PDFBookmark bookmark) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserBookmarksCollection(user.uid).doc(bookmark.id).set(bookmark.toMap());
+  }
+  
+  Future<void> updateBookmark(PDFBookmark bookmark) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserBookmarksCollection(user.uid).doc(bookmark.id).update(bookmark.toMap());
+  }
+  
+  Future<void> deleteBookmark(String documentId, String bookmarkId) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    await getUserBookmarksCollection(user.uid).doc(bookmarkId).delete();
+  }
+  
+  // PDF 파일 업로드/삭제
+  Future<String> uploadPDFFile(Uint8List fileBytes) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
+    
+    final filename = '${DateTime.now().millisecondsSinceEpoch}.pdf';
+    final ref = _storage.ref().child('${user.uid}/pdfs/$filename');
+    
+    // Uint8List를 사용하여 업로드
+    await ref.putData(fileBytes);
     return await ref.getDownloadURL();
   }
-
-  Future<void> deletePDFFile(String fileUrl) async {
-    final ref = _storage.refFromURL(fileUrl);
-    await ref.delete();
-  }
-
-  // 유틸리티 메서드
-  UserModel? _userCredentialToUserModel(UserCredential? credential) {
-    if (credential?.user == null) return null;
+  
+  Future<void> deletePDFFile(String filePath) async {
+    final user = _auth.currentUser;
+    if (user == null) throw Exception('User not logged in');
     
-    final user = credential!.user!;
-    return UserModel(
-      id: user.uid,
-      email: user.email ?? '',
-      displayName: user.displayName,
-      photoURL: user.photoURL,
-      createdAt: DateTime.now(),
-      updatedAt: DateTime.now(),
-      lastLoginAt: DateTime.now(),
-      settings: UserSettings.createDefault(),
-    );
+    // Firebase Storage 참조에서 파일 경로 추출
+    if (filePath.startsWith('gs://') || filePath.startsWith('http')) {
+      final ref = _storage.refFromURL(filePath);
+      await ref.delete();
+    } else {
+      // 상대 경로인 경우
+      final ref = _storage.ref().child(filePath);
+      await ref.delete();
+    }
   }
 } 
