@@ -1,46 +1,34 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:pdf_learner_v2/domain/models/user_model.dart';
 
-/// 인증 저장소 인터페이스
+/// 인증 관련 기능을 위한 리포지토리 인터페이스
 abstract class AuthRepository {
-  /// 사용자 인증 상태 스트림
-  Stream<UserModel?> get authStateChanges;
-
-  /// 이메일/비밀번호 인증
-  Future<UserModel?> signInWithEmailAndPassword(String email, String password);
-  Future<UserModel?> signUpWithEmailAndPassword(String email, String password);
+  /// 사용자 인증 상태 변화 추적
+  Stream<User?> authStateChanges();
+  
+  /// 현재 인증된 사용자 정보 가져오기
+  User? getCurrentUser();
+  
+  /// 이메일, 비밀번호로 로그인
+  Future<UserCredential> signIn(String email, String password);
+  
+  /// 이메일, 비밀번호로 회원가입
+  Future<UserCredential> signUp(String email, String password);
+  
+  /// 익명으로 로그인
+  Future<UserCredential> signInAnonymously();
+  
+  /// Google 계정으로 로그인
+  Future<UserCredential> signInWithGoogle();
+  
+  /// 로그아웃
   Future<void> signOut();
-
-  /// 사용자 정보 관리
-  Future<UserModel?> getCurrentUser();
-  Future<void> updateUser(UserModel user);
-  Future<void> deleteUser();
-
-  /// 비밀번호 관리
+  
+  /// 비밀번호 재설정 이메일 전송
   Future<void> resetPassword(String email);
-  Future<void> updatePassword(String newPassword);
-
-  /// 소셜 로그인
-  Future<UserModel?> signInWithGoogle();
-  Future<UserModel?> signInWithApple();
-  Future<UserModel?> signInWithFacebook();
-
-  /// 이메일 인증
-  Future<void> sendEmailVerification();
-  Future<void> verifyEmail();
-  Future<bool> isEmailVerified();
-  Future<void> changeEmail(String newEmail);
-
-  /// 프로필 관리
+  
+  /// 사용자 프로필 업데이트
   Future<void> updateProfile({
     String? displayName,
     String? photoURL,
   });
-
-  /// 계정 삭제
-  Future<void> deleteAccount();
-
-  /// 전화번호 인증
-  Future<void> verifyPhoneNumber(String phoneNumber);
-  Future<void> verifyPhoneCode(String verificationId, String code);
-} 
+}
