@@ -1,31 +1,54 @@
 import 'dart:io';
 import 'dart:typed_data';
-import '../models/pdf_document.dart';
-import '../models/pdf_bookmark.dart';
+import '../../domain/models/pdf_document.dart';
+import '../../domain/models/pdf_bookmark.dart';
 
-/// PDF 문서 관리를 위한 리포지토리 인터페이스
+/// PDF 문서와 북마크 관리를 위한 레포지토리 인터페이스
 abstract class PDFRepository {
-  /// 문서 관련 메서드
+  /// 모든 PDF 문서 조회
   Future<List<PDFDocument>> getDocuments();
-  Future<PDFDocument?> getDocument(String documentId);
+  
+  /// 특정 ID로 문서 조회
+  Future<PDFDocument?> getDocument(String id);
+  
+  /// 새 문서 생성
   Future<void> createDocument(PDFDocument document);
+  
+  /// 문서 정보 업데이트
   Future<void> updateDocument(PDFDocument document);
-  Future<void> deleteDocument(String documentId);
   
-  /// 북마크 관련 메서드
+  /// 문서 삭제
+  Future<void> deleteDocument(String id);
+  
+  /// 특정 문서의 모든 북마크 조회
   Future<List<PDFBookmark>> getBookmarks(String documentId);
-  Future<PDFBookmark?> getBookmark(String bookmarkId);
+  
+  /// 특정 북마크 조회
+  Future<PDFBookmark?> getBookmark(String documentId, String bookmarkId);
+  
+  /// 새 북마크 생성
   Future<void> createBookmark(PDFBookmark bookmark);
+  
+  /// 북마크 정보 업데이트
   Future<void> updateBookmark(PDFBookmark bookmark);
-  Future<void> deleteBookmark(String bookmarkId);
   
-  /// 파일 관련 메서드
-  Future<String> uploadPDFFile(String filePath, String fileName, {Uint8List? bytes});
-  Future<void> deletePDFFile(String fileUrl);
+  /// 북마크 삭제
+  Future<void> deleteBookmark(String documentId, String bookmarkId);
   
-  /// PDF 처리 관련 메서드
+  /// PDF 파일 업로드
+  Future<String> uploadPDFFile(Uint8List file);
+  
+  /// PDF 파일 삭제
+  Future<void> deletePDFFile(String filePath);
+  
+  /// PDF 페이지 수 가져오기
   Future<int> getPageCount(String filePath);
+  
+  /// PDF에서 텍스트 추출
   Future<String> extractText(String filePath, int pageNumber);
+  
+  /// PDF 메타데이터 조회
+  Future<Map<String, dynamic>> getMetadata(String filePath);
   
   /// 리소스 정리
   void dispose();
