@@ -12,32 +12,26 @@ import 'presentation/pages/signup_page.dart';
 import 'presentation/pages/reset_password_page.dart';
 import 'presentation/screens/splash_screen.dart';
 import 'package:get_it/get_it.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'presentation/app.dart';
 
 void main() async {
-  try {
-    WidgetsFlutterBinding.ensureInitialized();
-
-    // 환경 변수 로드
-    await dotenv.load(fileName: '.env');
-
-    // Firebase 초기화
-    await Firebase.initializeApp(
-      options: DefaultFirebaseOptions.currentPlatform,
-    );
-
-    // 의존성 주입 초기화
-    await configureDependencies();
-
-    runApp(const MyApp());
-  } catch (e) {
-    debugPrint('앱 초기화 중 오류 발생: $e');
-    // 에러 처리 로직 추가
-  }
+  WidgetsFlutterBinding.ensureInitialized();
+  
+  // Firebase 초기화
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  
+  // 의존성 주입 설정
+  await setupLocator();
+  
+  runApp(const App());
 }
 
 /// PDF Learner 앱
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class App extends StatelessWidget {
+  const App({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -53,8 +47,48 @@ class MyApp extends StatelessWidget {
       child: MaterialApp(
         title: 'PDF Learner',
         theme: ThemeData(
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+          colorScheme: ColorScheme.fromSeed(
+            seedColor: Colors.indigo,
+            brightness: Brightness.light,
+          ),
           useMaterial3: true,
+          cardTheme: CardTheme(
+            elevation: 0,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(16),
+            ),
+          ),
+          appBarTheme: const AppBarTheme(
+            centerTitle: false,
+            elevation: 0,
+            scrolledUnderElevation: 0,
+          ),
+          elevatedButtonTheme: ElevatedButtonThemeData(
+            style: ElevatedButton.styleFrom(
+              elevation: 0,
+              padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 24),
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(12),
+              ),
+            ),
+          ),
+          inputDecorationTheme: InputDecorationTheme(
+            filled: true,
+            fillColor: Colors.grey.shade100,
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12),
+              borderSide: BorderSide.none,
+            ),
+            enabledBorder: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12),
+              borderSide: BorderSide.none,
+            ),
+            focusedBorder: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12),
+              borderSide: BorderSide(color: Colors.indigo.shade200),
+            ),
+            contentPadding: const EdgeInsets.all(16),
+          ),
         ),
         home: const SplashScreen(),
         routes: {
