@@ -2,66 +2,131 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../viewmodels/auth_view_model.dart';
 import '../pages/home_page.dart';
-import '../pages/login_page.dart';
 
 class SplashScreen extends StatefulWidget {
-  const SplashScreen({super.key});
+  const SplashScreen({Key? key}) : super(key: key);
 
   @override
   State<SplashScreen> createState() => _SplashScreenState();
 }
 
-class _SplashScreenState extends State<SplashScreen> {
+class _SplashScreenState extends State<SplashScreen> with SingleTickerProviderStateMixin {
+  late AnimationController _animationController;
+  late Animation<double> _fadeInAnimation;
+
   @override
   void initState() {
     super.initState();
-    // 초기화 후 상태 확인
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      final authViewModel = Provider.of<AuthViewModel>(context, listen: false);
-      _checkAuthState(authViewModel);
+    
+    // 애니메이션 컨트롤러 설정
+    _animationController = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 1500),
+    );
+    
+    _fadeInAnimation = CurvedAnimation(
+      parent: _animationController,
+      curve: Curves.easeIn,
+    );
+    
+    // 애니메이션 시작
+    _animationController.forward();
+    
+    // 2초 후 자동 익명 로그인하고 홈 화면으로 이동
+    Future.delayed(const Duration(seconds: 2), () {
+      _signInAnonymously();
     });
   }
+  
+  @override
+  void dispose() {
+    _animationController.dispose();
+    super.dispose();
+  }
 
-  void _checkAuthState(AuthViewModel authViewModel) {
-    // 인증 상태에 따라 페이지 이동
-    if (authViewModel.isAuthenticated) {
-      Navigator.of(context).pushReplacementNamed('/home');
-    } else if (authViewModel.isUnauthenticated) {
-      Navigator.of(context).pushReplacementNamed('/login');
-    } else if (authViewModel.hasError) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text(authViewModel.error ?? '인증 오류가 발생했습니다')),
+  // 익명 로그인 후 홈 화면으로 이동
+  Future<void> _signInAnonymously() async {
+    final authViewModel = Provider.of<AuthViewModel>(context, listen: false);
+    await authViewModel.signInAnonymously();
+    
+    if (mounted) {
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(
+          builder: (context) => const HomePage(),
+        ),
       );
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<AuthViewModel>(
-      builder: (context, authViewModel, child) {
-        // 상태가 변경되면 다시 확인
-        if (authViewModel.isAuthenticated) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            Navigator.of(context).pushReplacementNamed('/home');
-          });
-        } else if (authViewModel.isUnauthenticated) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            Navigator.of(context).pushReplacementNamed('/login');
-          });
-        } else if (authViewModel.hasError) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(content: Text(authViewModel.error ?? '인증 오류가 발생했습니다')),
-            );
-          });
-        }
-
-        return const Scaffold(
-          body: Center(
-            child: CircularProgressIndicator(),
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: Container(
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              Colors.white,
+              const Color(0xFFF5F7FA),
+            ],
           ),
-        );
-      },
+        ),
+        child: SafeArea(
+          child: Center(
+            child: FadeTransition(
+              opacity: _fadeInAnimation,
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  // 앱 아이콘
+                  Container(
+                    padding: const EdgeInsets.all(24),
+                    decoration: BoxDecoration(
+                      gradient: LinearGradient(
+                        colors: [
+                          const Color(0xFF5D5FEF),
+                          const Color(0xFF3D6AFF),
+                        ],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      ),
+                      borderRadius: BorderRadius.circular(24),
+                    ),
+                    child: const Icon(
+                      Icons.picture_as_pdf_rounded,
+                      size: 72,
+                      color: Colors.white,
+                    ),
+                  ),
+                  const SizedBox(height: 24.0),
+                  Text(
+                    'PDF 학습기',
+                    style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                      fontWeight: FontWeight.bold,
+                      color: const Color(0xFF5D5FEF),
+                    ),
+                  ),
+                  const SizedBox(height: 8),
+                  Text(
+                    'PDF 문서를 효율적으로 관리하고 학습하세요',
+                    style: TextStyle(
+                      fontSize: 14,
+                      color: Colors.grey.shade600,
+                    ),
+                    textAlign: TextAlign.center,
+                  ),
+                  const SizedBox(height: 40),
+                  CircularProgressIndicator(
+                    valueColor: AlwaysStoppedAnimation<Color>(const Color(0xFF5D5FEF)),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
+      ),
     );
   }
 } 
